ANDROID: sched: fair: Bypass energy-aware wakeup for prefer-idle tasks
Use the upstream slow path to find an idle cpu for prefer-idle tasks.
This slow-path is actually faster than the EAS path we are currently
going through (select_energy*) which is really slow.

No performance degradation is seen with this and it reduces the delta
quite a bit between upstream and out of tree code.

It's not clear yet if using the mainline slow path task placement when
a task has the schedtune attribute prefer_idle=1 is the right thing to
do for products. Put the option to disable this behind a sched feature
so we can try out both options.

Change-Id: Icf762a101c92c0e3f9e61df0370247fa15455581
Signed-off-by: Joel Fernandes <joelaf@google.com>
(refactored for 4.14 version)
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
idkwhoiam322: Port to 4.4
Cherry-picked from: https://android.googlesource.com/kernel/msm.git/+/c0ff131c88f68e4985793663144b6f9cf77be9d3
Signed-off-by: idkwhoiam322 <idkwhoiam322@raphielgang.org>